axf/requirements.txt
Django==2.0.7
django-cors-headers==3.1.0
django-filter==2.2.0
django-redis==4.10.0
django-stubs==1.1.0
djangorestframework==3.10.3
djangorestframework-stubs==1.0.0
mypy==0.720
mypy-extensions==0.4.1
Pillow==6.1.0
PyMySQL==0.9.3
pytz==2019.2
redis==3.3.8
typed-ast==1.4.0
typing-extensions==3.7.4
